fix(ash-events): capture actor id independently of actor shape
`ApiEvent.user_id` (via AshEvents' `persist_actor_primary_key`) only
populates when the actor is an exact `%ServiceRadar.Identity.User{}`
struct. Both real actor-construction paths -- `set_ash_actor` (the
`:ash_json_api` router pipeline) and the `Ash.Scope.ToOpts`
implementation for `Scope` (the LiveView path) -- build a plain map
instead, so on real requests through either transport `user_id` was
silently nil and actor attribution was lost for the audit log's core
use case.

`StampEventSource` now also reads `context.actor` (the third,
`%Ash.Resource.Change.Context{}`, argument to `change/3`) and stamps
`metadata["actor_id"]` with the actor's `:id`, regardless of whether
the actor is a struct or a map. `AuditHistory`'s adapter now prefers
`user_id` when present (struct actors, e.g. system/background actors,
stay covered) and falls back to `metadata["actor_id"]` otherwise.

Adds a test using the exact plain-map actor shape both real
call sites produce (not a hand-built `%User{}` struct, which was
masking this gap in the existing coverage), and fixes the
`api_events.version` migration column to `:integer` to match the
AshEvents attribute type it backs (`:bigint` was a pre-ship typo,
never deployed).

Verified against a fresh srql-fixtures scratch database: both
`stateful_alert_rule_events_test.exs` and
`audit_history_ash_events_test.exs` pass (14/14), and the new test
was confirmed to fail against the pre-fix code with
`event.metadata["actor_id"] == nil`.

+++ elixir/serviceradar_core/priv/repo/migrations/20260909100000_add_api_event_log.exs
@@ -29,7 +29,7 @@ defmodule ServiceRadar.Repo.Migrations.AddApiEventLog do
     create_if_not_exists table(:api_events, primary_key: false, prefix: @prefix) do
       add(:id, :uuid, null: false, default: fragment("uuid_generate_v7()"), primary_key: true)
       add(:record_id, :uuid, null: false)
-      add(:version, :bigint, null: false, default: 1)
+      add(:version, :integer, null: false, default: 1)
       add(:occurred_at, :utc_datetime_usec, null: false, default: fragment("now()"))
       add(:resource, :text, null: false)
       add(:action, :text, null: false)
